+++ tags_widthratio.go
@@ -80,7 +80,7 @@ func (node *tagWidthratioNode) Execute(ctx *ExecutionContext, writer TemplateWri
 			return err
 		}
 	} else {
-		ctx.Private[node.ctxName] = value
+		ctx.Private.Set(node.ctxName, value)
 	}
 
 	return nil